+++ cterasdk/common/__init__.py
@@ -1,7 +1,7 @@
 from .item import Item  # noqa: E402, F401
 from .object import Object, Device, delete_attrs  # noqa: E402, F401
-from .datetime_utils import DateTimeUtils, from_iso_format  # noqa: E402, F401
-from .utils import merge, union, parse_base_object_ref, convert_size, df_military_time, DataUnit, parse_to_ipaddress, \
+from .datetime_utils import DateTimeUtils  # noqa: E402, F401
+from .utils import merge, union, parse_base_object_ref, convert_size, df_military_time, DataUnit, \
                    utf8_decode, utf8_encode, Version  # noqa: E402, F401
 from .types import PolicyRule, PolicyRuleConverter, StringCriteriaBuilder, IntegerCriteriaBuilder, DateTimeCriteriaBuilder, \
                    PredefinedListCriteriaBuilder, CustomListCriteriaBuilder, ThrottlingRuleBuilder, ThrottlingRule, FilterBackupSet, \