+++ cterasdk/common/datetime_utils.py
@@ -19,14 +19,3 @@ class DateTimeUtils():
         elif isinstance(expiration, datetime.date):
             expiration_date = expiration
         return expiration_date  # pylint: disable=possibly-used-before-assignment
-
-
-def from_iso_format(time):
-    """
-    Parse datetime object from ISO 8601 format
-
-    :param str time: Timestamp
-    :returns: Datetime object
-    :rtype: datetime.datetime
-    """
-    return datetime.datetime.fromisoformat(time)